+++ packages/react-test-renderer/src/ReactTestRendererAct.js
@@ -6,7 +6,7 @@
  *
  * @flow
  */
-import type {Thenable} from 'react-reconciler/src/ReactFiberScheduler';
+import type {Thenable} from 'react-reconciler/src/ReactFiberWorkLoop';
 
 import {
   batchedUpdates,
